requirements.txt
googlesearch-python
Flask
Gunicorn
requests